refactor: isolate build and diff side orchestration
